Add auto to palette list (not done)
