+++ MahMaterialDragablzMashUp/PaletteSelectorViewModel.cs
@@ -14,10 +14,6 @@ namespace MahMaterialDragablzMashUp
             ApplyPrimaryCommand = new DelegateCommand<Swatch>(ApplyPrimary, new Predicate<Swatch>(s => s != Primary));
             ApplyAccentCommand = new DelegateCommand<Swatch>(ApplyAccent, new Predicate<Swatch>(s => s != Accent));
             ToggleBaseCommand = new DelegateCommand<bool?>(SetLightDark);
-		    ApplyPrimary(this.AutoPrimary);
-            ApplyAccent(this.AutoAccent);
-            ApplyPrimaryCommand.RaiseCanExecuteChanged();
-            ApplyAccentCommand.RaiseCanExecuteChanged();
         }
 
         public ICommand ToggleBaseCommand { get; }
