+++ MainDemo.Wpf/PaletteSelectorViewModel.cs
@@ -16,14 +16,10 @@ namespace MaterialDesignColors.WpfExample
     {
         public PaletteSelectorViewModel()
         {
-            Swatches = new SwatchesProvider().Swatches;
+            Swatches = new SwatchesProvider(true).Swatches;
             ApplyPrimaryCommand = new DelegateCommand<Swatch>(ApplyPrimary, new Predicate<Swatch>(s => s != Primary));
             ApplyAccentCommand = new DelegateCommand<Swatch>(ApplyAccent, new Predicate<Swatch>(s => s != Accent));
             ToggleBaseCommand = new DelegateCommand<bool?>(SetLightDark);
-            ApplyPrimary(this.AutoPrimary);
-            ApplyAccent(this.AutoAccent);
-            ApplyPrimaryCommand.RaiseCanExecuteChanged();
-            ApplyAccentCommand.RaiseCanExecuteChanged();
         }
 
         public ICommand ToggleBaseCommand { get; }
